requirements.txt
fastapi[all]==0.121.2
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0

sqlalchemy==2.0.44
asyncpg==0.30.0
greenlet==3.2.4
alembic==1.17.2

python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
boto3==1.40.74

structlog==25.5.0
python-json-logger==4.0.0
pyotp==2.9.0
qrcode[pil]==8.2

redis==7.0.1
celery==5.5.3

pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1
jinja2==3.1.6
